Move code to transaction_batch_insert_builder.go
